INFINIBOX-7019: introduced partition key and added automatic discovery and duplication of key fields.
CQL looks similar to SQL but it is very limited.
There are 3 types of fields: keys, simple fields and indexed fields (secondary index).
Onley keys and indexed fields can be used in WHERE clause. Although AND relationship is supported keys and indexed fileds cannot be combined.
ORDER BY statment supports keys only. Morevoer keys must follow thier order in table definition. The first key must be limited using = or IN statement.

So, here is what was implemented.
1. We parse prepared statements when creating DB schema
2. We add missing indexes
3. We create "duplicate" key fields to enable order by.

We still do not create special search tables to support several different sequences of order by although the environment is ready. We are planning to implement this soon.