Added support of per-record and aggregation functions.
Per-record functions are functions that are executed on each selected record separately. For example upper(), lower(), length().
Aggregation functions are those that calculate value using specific field of all selected records (e.g. max(), min(), sum(), avg()).

This commit contains also fixes of integration test that did not work if Cassandra is not running. Fixing this problem exposed bug in production
code that was caused by technique that we use for storing the global instance of Cassandra daemon (currently for tests only).