INFINIBOX-7114: limited cascading implementation
Limited implementaiton of cascades. Cascades are fully supported when persisting and deleting entitites.
However orphanRemoval is not supported right now. Comment that describes how to implement this feature can be
found in CqlPersistor.convertValue().
Cascade types MERGE and REFRESH are not supported. Fetch type is ignored.

JPA defines 3 different things that are merged into one in Mestor: cascade type, fetch type and attribute "orphanRemoval"
supported by @OneToMany and @ManyToMany. All these parameters are parsed and "attached" for FieldMetadata as a set of
Mestor's "CascadeOption". Not all CascadeOptions are fully supported.

    PERSIST: done
    MERGE:   -
    REMOVE:  done except orphanRemoval
    REFRESH: -
    DETACH:  -
    FETCH (JPA's fetch type EAGER): not implemented, howerver it may just cause some performance issues but not functionality problems.
    ORPHAN_REMOVAL: -